refactor: everytime 모듈을 lecture 패키지로 이동
