+++ backend/src/main/java/unischedule/external/OpenAiClient.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Component;
 import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.reactive.function.client.WebClient;
 import reactor.core.publisher.Mono;
-import unischedule.everytime.dto.TimetableDetailDto;
+import unischedule.lecture.everytime.dto.TimetableDetailDto;
 import unischedule.exception.ExternalApiException;
 import unischedule.external.dto.OpenAiChatCompletionResponseDto;
