+++ backend/src/main/java/unischedule/lecture/everytime/dto/TimetableDetailDto.java
@@ -1,4 +1,11 @@
-package unischedule.everytime.dto;
+package unischedule.lecture.everytime.dto;
+
+import jakarta.validation.Valid;
+import jakarta.validation.constraints.Max;
+import jakarta.validation.constraints.Min;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotEmpty;
+import jakarta.validation.constraints.NotNull;
 
 import java.time.LocalTime;
 import java.util.List;
@@ -6,6 +13,8 @@ import java.util.List;
 public record TimetableDetailDto(
         String year,
         String semester,
+
+        @Valid
         List<Subject> subjects
 ) {
 
@@ -14,9 +23,16 @@ public record TimetableDetailDto(
     }
 
     public record Subject(
+            @NotBlank(message = "강의명은 필수입니다")
             String name,
+            
+            @NotBlank(message = "교수명은 필수입니다")
             String professor,
+
             Integer credit,
+            
+            @NotEmpty(message = "강의 시간은 최소 1개 이상이어야 합니다")
+            @Valid
             List<Time> times
     ) {
 
@@ -26,9 +42,17 @@ public record TimetableDetailDto(
         }
 
         public record Time(
+                @NotNull(message = "요일은 필수입니다")
+                @Min(value = 1, message = "요일은 1(월)부터 7(일) 사이여야 합니다")
+                @Max(value = 7, message = "요일은 1(월)부터 7(일) 사이여야 합니다")
                 Integer dayOfWeek,
+                
+                @NotNull(message = "시작 시간은 필수입니다")
                 LocalTime startTime,
+                
+                @NotNull(message = "종료 시간은 필수입니다")
                 LocalTime endTime,
+
                 String place
         ) {
 
@@ -39,3 +63,4 @@ public record TimetableDetailDto(
         }
     }
 }
+
